+++ src/main/java/com/ghgande/j2mod/modbus/facade/AbstractModbusMaster.java
@@ -46,6 +46,7 @@ abstract public class AbstractModbusMaster {
     private ReadMultipleRegistersRequest readMultipleRegistersRequest;
     private WriteSingleRegisterRequest writeSingleRegisterRequest;
     private WriteMultipleRegistersRequest writeMultipleRegistersRequest;
+    private MaskWriteRegisterRequest maskWriteRegisterRequest;
     protected int timeout = Modbus.DEFAULT_TIMEOUT;
 
     /**
@@ -245,10 +246,12 @@ abstract public class AbstractModbusMaster {
      * @param register a <tt>Register</tt> holding the value of the register
      *                 to be written.
      *
+     * @return the value of the register as returned from the slave.
+     *
      * @throws ModbusException if an I/O error, a slave exception or
      *                         a transaction error occurs.
      */
-    public void writeSingleRegister(int unitId, int ref, Register register) throws ModbusException {
+    public int writeSingleRegister(int unitId, int ref, Register register) throws ModbusException {
         checkTransaction();
         if (writeSingleRegisterRequest == null) {
             writeSingleRegisterRequest = new WriteSingleRegisterRequest();
@@ -258,6 +261,7 @@ abstract public class AbstractModbusMaster {
         writeSingleRegisterRequest.setRegister(register);
         transaction.setRequest(writeSingleRegisterRequest);
         transaction.execute();
+        return ((WriteSingleRegisterResponse) getAndCheckResponse()).getRegisterValue();
     }
 
     /**
@@ -268,10 +272,12 @@ abstract public class AbstractModbusMaster {
      * @param registers a <tt>Register[]</tt> holding the values of
      *                  the registers to be written.
      *
+     * @return the number of registers that have been written.
+     *
      * @throws ModbusException if an I/O error, a slave exception or
      *                         a transaction error occurs.
      */
-    public void writeMultipleRegisters(int unitId, int ref, Register[] registers) throws ModbusException {
+    public int writeMultipleRegisters(int unitId, int ref, Register[] registers) throws ModbusException {
         checkTransaction();
         if (writeMultipleRegistersRequest == null) {
             writeMultipleRegistersRequest = new WriteMultipleRegistersRequest();
@@ -281,6 +287,38 @@ abstract public class AbstractModbusMaster {
         writeMultipleRegistersRequest.setRegisters(registers);
         transaction.setRequest(writeMultipleRegistersRequest);
         transaction.execute();
+        return ((WriteMultipleRegistersResponse) transaction.getResponse()).getWordCount();
+    }
+
+    /**
+     * Mask write a single register to the slave.
+     *
+     * @param unitId    the slave unit id.
+     * @param ref       the offset of the register to start writing to.
+     * @param andMask   AND mask.
+     * @param orMask    OR mask.
+     *
+     * @return true if success, i.e. response data equals to request data, false otherwise.
+     *
+     * @throws ModbusException if an I/O error, a slave exception or
+     *                         a transaction error occurs.
+     */
+    public boolean maskWriteRegister(int unitId, int ref, int andMask, int orMask) throws ModbusException {
+        checkTransaction();
+        if (maskWriteRegisterRequest == null) {
+            maskWriteRegisterRequest = new MaskWriteRegisterRequest();
+        }
+        maskWriteRegisterRequest.setUnitID(unitId);
+        maskWriteRegisterRequest.setReference(ref);
+        maskWriteRegisterRequest.setAndMask(andMask);
+        maskWriteRegisterRequest.setOrMask(orMask);
+        transaction.setRequest(maskWriteRegisterRequest);
+        transaction.execute();
+
+        MaskWriteRegisterResponse response = (MaskWriteRegisterResponse) getAndCheckResponse();
+        return response.getReference() == maskWriteRegisterRequest.getReference() &&
+               response.getAndMask() == maskWriteRegisterRequest.getAndMask() &&
+               response.getOrMask() == maskWriteRegisterRequest.getOrMask();
     }
 
     /**
@@ -395,11 +433,13 @@ abstract public class AbstractModbusMaster {
      * @param register a <tt>Register</tt> holding the value of the register
      *                 to be written.
      *
+     * @return the value of the register as returned from the slave.
+     *
      * @throws ModbusException if an I/O error, a slave exception or
      *                         a transaction error occurs.
      */
-    public void writeSingleRegister(int ref, Register register) throws ModbusException {
-        writeSingleRegister(DEFAULT_UNIT_ID, ref, register);
+    public int writeSingleRegister(int ref, Register register) throws ModbusException {
+        return writeSingleRegister(DEFAULT_UNIT_ID, ref, register);
     }
 
     /**
@@ -409,11 +449,29 @@ abstract public class AbstractModbusMaster {
      * @param registers a <tt>Register[]</tt> holding the values of
      *                  the registers to be written.
      *
+     * @return the number of registers that have been written.
+     *
      * @throws ModbusException if an I/O error, a slave exception or
      *                         a transaction error occurs.
      */
-    public void writeMultipleRegisters(int ref, Register[] registers) throws ModbusException {
-        writeMultipleRegisters(DEFAULT_UNIT_ID, ref, registers);
+    public int writeMultipleRegisters(int ref, Register[] registers) throws ModbusException {
+        return writeMultipleRegisters(DEFAULT_UNIT_ID, ref, registers);
+    }
+
+    /**
+     * Mask write a single register to the slave.
+     *
+     * @param ref       the offset of the register to start writing to.
+     * @param andMask   AND mask.
+     * @param orMask    OR mask.
+     *
+     * @return true if success, i.e. response data equals to request data, false otherwise.
+     *
+     * @throws ModbusException if an I/O error, a slave exception or
+     *                         a transaction error occurs.
+     */
+    public boolean maskWriteRegister(int ref, int andMask, int orMask) throws ModbusException {
+        return maskWriteRegister(DEFAULT_UNIT_ID, ref, andMask, orMask);
     }
 
     /**